printf test_mix solved

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bhocsak <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 13:20:16 by bhocsak           #+#    #+#             */
-/*   Updated: 2024/05/07 16:20:57 by bhocsak          ###   ########.fr       */
+/*   Updated: 2024/05/30 17:48:57 by bhocsak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,6 @@ int		ft_putnbr_pf(int n);
 int		ft_putstr_pf(char *str);
 int		ft_putptr(void *ptr);
 int		ft_putuint(unsigned int n);
-int		ft_puthex(unsigned long n, const char base);
+int		ft_puthex(unsigned int n, const char base);
 
 #endif
